# Conflicts:
#	.github/dependabot.yml
#	.github/dependencies.json
#	.github/scripts/dependency-report.mjs
#	.github/scripts/dependency-report.test.mjs
#	.github/scripts/release-showcase.mjs
#	.github/scripts/release-showcase.test.mjs
#	.github/scripts/repository-check.mjs
#	.github/scripts/step-summary.mjs
#	.github/scripts/step-summary.test.mjs
#	.github/workflows/build.yml
#	.github/workflows/dependencies.yml
#	.github/workflows/firmware.yml
#	.github/workflows/host.yml
#	.github/workflows/release.yml
#	.github/workflows/repository-health.yml
#	.github/workflows/virtual-board.yml
#	README.md
#	Tools/Controller/internal/control/history_test.go
#	Tools/Controller/internal/native/frame_test.go
#	Tools/Controller/internal/native/rf_test.go
#	docs/CI-CD-and-Releases.md

+++ Tools/VirtualBoard/src/virtual_board.cpp
@@ -1485,7 +1485,7 @@ wire::Frame VirtualBoard::remotesFrame(std::uint8_t sequence,
           remotes_.begin(), remotes_.end(),
           [](const RemoteEntry &entry) { return entry.used; }));
   std::vector<std::uint8_t> payload{1, total, 0xFF, 0};
-  std::uint8_t scan = cursor;
+  std::size_t scan = cursor;
   while (scan < remotes_.size() && payload[3] < 3) {
     const RemoteEntry &entry = remotes_[scan];
     if (entry.used) {
@@ -1503,7 +1503,7 @@ wire::Frame VirtualBoard::remotesFrame(std::uint8_t sequence,
   }
   while (scan < remotes_.size()) {
     if (remotes_[scan].used) {
-      payload[2] = scan;
+      payload[2] = static_cast<std::uint8_t>(scan);
       break;
     }
     ++scan;
